chore(pass-authenticators-webauthn): refactored mock structure, and split tests
