+++ authenticators/webauthn/src/mock.rs
@@ -1,21 +1,122 @@
-use super::*;
+//! Test environment for pass webauthn.
 
-use codec::Decode;
-use frame::deps::sp_runtime::traits::TrailingZeroInput;
-use frame_system::pallet_prelude::BlockNumberFor;
+use frame::{
+    hashing,
+    testing_prelude::*,
+    traits::{ConstU32, EqualPrivilegeOnly},
+};
 use futures::executor::block_on;
-
-use passkey_authenticator::{Authenticator, MockUserValidationMethod};
+use pallet_pass::AddressGenerator;
+use passkey_authenticator::MockUserValidationMethod;
 use passkey_client::{Client, DefaultClientData};
-use passkey_types::{ctap2::Aaguid, webauthn::*, Bytes, Passkey};
-
-use hashing::blake2_256;
-use traits_authn::{AuthorityId, Challenger, HashedUserId};
+use passkey_types::ctap2::Aaguid;
+use passkey_types::{Bytes, Passkey};
+use passkey_types::webauthn::{AttestationConveyancePreference, AttestationStatementFormatIdentifiers, CredentialCreationOptions, CredentialRequestOptions, PublicKeyCredentialCreationOptions, PublicKeyCredentialDescriptor, PublicKeyCredentialParameters, PublicKeyCredentialRequestOptions, PublicKeyCredentialRpEntity, PublicKeyCredentialType, PublicKeyCredentialUserEntity, UserVerificationRequirement};
+use traits_authn::{util::AuthorityFromPalletId, Challenger, ExtrinsicContext, HashedUserId};
 use url_evil::Url;
+use crate::{AssertionMeta, Authenticator, DEREncodedPublicKey};
 
-use crate::{AssertionMeta, DEREncodedPublicKey};
+#[frame_construct_runtime]
+pub mod runtime {
+    #[runtime::runtime]
+    #[runtime::derive(
+        RuntimeCall,
+        RuntimeEvent,
+        RuntimeError,
+        RuntimeOrigin,
+        RuntimeTask,
+        RuntimeHoldReason,
+        RuntimeFreezeReason
+    )]
+    pub struct Test;
 
-use super::{BlockChallenger, Test};
+    #[runtime::pallet_index(0)]
+    pub type System = frame_system;
+    #[runtime::pallet_index(1)]
+    pub type Scheduler = pallet_scheduler;
+    #[runtime::pallet_index(2)]
+    pub type Pass = pallet_pass;
+
+    #[runtime::pallet_index(10)]
+    pub type Balances = pallet_balances;
+}
+
+pub type Block = frame_system::mocking::MockBlock<Test>;
+pub type AccountId = <Test as frame_system::Config>::AccountId;
+
+#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
+impl frame_system::Config for Test {
+    type Block = Block;
+    type AccountData = pallet_balances::AccountData<AccountId>;
+}
+
+#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
+impl pallet_balances::Config for Test {
+    type AccountStore = System;
+}
+
+parameter_types! {
+    pub MaxWeight: weights::Weight = weights::Weight::MAX;
+}
+
+impl pallet_scheduler::Config for Test {
+    type RuntimeEvent = RuntimeEvent;
+    type RuntimeOrigin = RuntimeOrigin;
+    type PalletsOrigin = OriginCaller;
+    type RuntimeCall = RuntimeCall;
+    type MaximumWeight = MaxWeight;
+    type ScheduleOrigin = EnsureRoot<AccountId>;
+    type OriginPrivilegeCmp = EqualPrivilegeOnly;
+    type MaxScheduledPerBlock = ConstU32<256>;
+    type WeightInfo = ();
+    type Preimages = ();
+    type BlockNumberProvider = System;
+}
+
+parameter_types! {
+  pub PassPalletId: PalletId = PalletId(*b"pass_web");
+  pub NeverPays: Option<pallet_pass::DepositInformation<Test>> = None;
+}
+
+pub type AuthorityId = AuthorityFromPalletId<PassPalletId>;
+
+pub struct BlockChallenger;
+
+impl Challenger for BlockChallenger {
+    type Context = BlockNumberFor<Test>;
+
+    fn generate(ctx: &Self::Context, xtc: &impl ExtrinsicContext) -> traits_authn::Challenge {
+        <Test as frame_system::Config>::Hashing::hash(&((ctx, xtc.as_ref()).encode())).0
+    }
+}
+
+pub struct SumAddressGenerator;
+impl AddressGenerator<Test, ()> for SumAddressGenerator {
+    fn generate_address(id: HashedUserId) -> AccountId {
+        id.iter()
+            .map(|b| *b as u64)
+            .reduce(Saturating::saturating_add)
+            .unwrap()
+    }
+}
+
+impl pallet_pass::Config for Test {
+    type RuntimeEvent = RuntimeEvent;
+    type PalletsOrigin = OriginCaller;
+    type RuntimeCall = RuntimeCall;
+    type WeightInfo = ();
+    type RegisterOrigin = EnsureRootWithSuccess<Self::AccountId, ConstU64<0>>;
+    type AddressGenerator = SumAddressGenerator;
+    type Balances = Balances;
+    type Authenticator = Authenticator<BlockChallenger, AuthorityId>;
+    type Scheduler = Scheduler;
+    type BlockNumberProvider = System;
+    type RegistrarConsideration = ();
+    type DeviceConsideration = ();
+    type SessionKeyConsideration = ();
+    type PalletId = PassPalletId;
+    type MaxSessionDuration = ConstU64<10>;
+}
 
 pub struct WebAuthnClient {
     origin: Url,
@@ -25,7 +126,7 @@ pub struct WebAuthnClient {
 impl WebAuthnClient {
     pub fn new(origin: &'static str, times: usize) -> Self {
         // Create Authenticator
-        let authenticator = Authenticator::new(
+        let authenticator = passkey_authenticator::Authenticator::new(
             Aaguid::new_empty(),
             None,
             MockUserValidationMethod::verified_user(times),
@@ -73,7 +174,7 @@ impl WebAuthnClient {
             creation_options,
             DefaultClientData,
         ))
-        .map_err(|_| ())?;
+            .map_err(|_| ())?;
 
         let public_key: DEREncodedPublicKey = result
             .response
@@ -121,7 +222,7 @@ impl WebAuthnClient {
             request_options,
             DefaultClientData,
         ))
-        .map_err(|_| ())?;
+            .map_err(|_| ())?;
 
         // Extracting required fields
         let user_handle = result
@@ -141,7 +242,7 @@ impl WebAuthnClient {
         user_id: HashedUserId,
         context: BlockNumberFor<Test>,
         xtc: &impl ExtrinsicContext,
-        authority_id: AuthorityId,
+        authority_id: traits_authn::AuthorityId,
     ) -> (Vec<u8>, crate::Attestation<BlockNumberFor<Test>>) {
         let challenge = BlockChallenger::generate(&context, xtc);
 
@@ -169,7 +270,7 @@ impl WebAuthnClient {
         credential_id: impl Into<Bytes>,
         context: BlockNumberFor<Test>,
         xtc: &impl ExtrinsicContext,
-        authority_id: AuthorityId,
+        authority_id: traits_authn::AuthorityId,
     ) -> crate::Assertion<BlockNumberFor<Test>> {
         let challenge = BlockChallenger::generate(&context, xtc);
 
@@ -188,4 +289,19 @@ impl WebAuthnClient {
             signature,
         }
     }
+}
+
+pub struct TestExt(pub TestExternalities, pub WebAuthnClient);
+impl TestExt {
+    pub fn execute_with<R>(&mut self, execute: impl FnOnce(&mut WebAuthnClient) -> R) -> R {
+        self.0.execute_with(|| execute(&mut self.1))
+    }
+}
+
+pub fn new_test_ext(times: usize) -> TestExt {
+    let mut t = TestExternalities::default();
+    t.execute_with(|| {
+        System::set_block_number(1);
+    });
+    TestExt(t, WebAuthnClient::new("https://pass_web.pass.int", times))
 }
